+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v1.0.3
 ======
+* Add display outlines on preview
 * UI and command line implementation of image optimization
 * Change zoom slider range (min: 10, max: 1000)
 * Added automatically refreshed preview